refactor: restructure file system modules and consolidate functionality
- Replace File.js with new FS.js module for better organization
- Enhance FileObject.js with expanded functionality
- Update DirectoryObject.js with improved methods
- Consolidate type definitions and remove deprecated File types
- Remove outdated FileSystem examples
- Update package version and main exports
- Improve code organization and reduce overall codebase size
