--- examples/FileSystem/package.json
@@ -1,9 +1,0 @@
-{
-  "type": "module",
-  "dependencies": {
-    "@gesslar/toolkit": "file:../.."
-  },
-  "devDependencies": {
-    "@types/node": "^24.2.1"
-  }
-}
