+++ src/lib/FileObject.js
@@ -4,11 +4,16 @@
  * resolution and existence checks.
  */
 
+import JSON5 from "json5"
+import fs from "node:fs/promises"
 import path from "node:path"
 import util from "node:util"
+import YAML from "yaml"
 
 import DirectoryObject from "./DirectoryObject.js"
-import File from "./File.js"
+import FS from "./FS.js"
+import Sass from "./Sass.js"
+import Valid from "./Valid.js"
 
 /**
  * FileObject encapsulates metadata and operations for a file, including path
@@ -26,7 +31,7 @@ import File from "./File.js"
  * @property {Promise<boolean>} exists - Whether the file exists (async)
  */
 
-export default class FileObject {
+export default class FileObject extends FS {
   /**
    * @type {object}
    * @private
@@ -59,9 +64,11 @@ export default class FileObject {
    * @param {DirectoryObject|string|null} [directory] - The parent directory (object or string)
    */
   constructor(fileName, directory=null) {
-    const fixedFile = File.fixSlashes(fileName)
+    super(fileName, directory)
 
-    const {dir,base,ext} = File.deconstructFilenameToParts(fixedFile)
+    const fixedFile = FS.fixSlashes(fileName)
+
+    const {dir,base,ext} = this.#deconstructFilenameToParts(fixedFile)
 
     if(!directory)
       directory = new DirectoryObject(dir)
@@ -71,7 +78,7 @@ export default class FileObject {
       : path.resolve(directory?.path ?? ".", fixedFile)
 
     const resolved = final
-    const fileUri = File.pathToUri(resolved)
+    const fileUri = FS.pathToUri(resolved)
 
     this.#meta.supplied = fixedFile
     this.#meta.path = resolved
@@ -80,7 +87,7 @@ export default class FileObject {
     this.#meta.extension = ext
     this.#meta.module = path.basename(this.supplied, this.extension)
 
-    const {dir: newDir} = File.deconstructFilenameToParts(this.path)
+    const {dir: newDir} = this.#deconstructFilenameToParts(this.path)
 
     this.#meta.directory = new DirectoryObject(newDir)
 
@@ -127,7 +134,7 @@ export default class FileObject {
    * @returns {Promise<boolean>} - A Promise that resolves to true or false
    */
   get exists() {
-    return File.fileExists(this)
+    return this.#fileExists()
   }
 
   /**
@@ -218,5 +225,162 @@ export default class FileObject {
    */
   get directory() {
     return this.#meta.directory
+  }
+
+  /**
+   * Check if a file can be read. Returns true if the file can be read, false
+   *
+   * @returns {Promise<boolean>} Whether the file can be read
+   */
+  async canRead() {
+    try {
+      await fs.access(this.path, fs.constants.R_OK)
+
+      return true
+    } catch(_) {
+      return false
+    }
+  }
+
+  /**
+   * Check if a file can be written. Returns true if the file can be written,
+   *
+   * @returns {Promise<boolean>} Whether the file can be written
+   */
+  async canWrite() {
+    try {
+      await fs.access(this.path, fs.constants.W_OK)
+
+      return true
+    } catch(_) {
+      return false
+    }
+  }
+
+  /**
+   * Check if a file exists
+   *
+   * @returns {Promise<boolean>} Whether the file exists
+   */
+  async #fileExists() {
+    try {
+      await fs.access(this.path, fs.constants.F_OK)
+
+      return true
+    } catch(_) {
+      return false
+    }
+  }
+
+  /**
+   * Determines the size of a file.
+   *
+   * @returns {Promise<number?>} - The size of the file or null, if it doesn't exist.
+   */
+  async size() {
+    try {
+      const stat = await fs.stat(this.path)
+
+      return stat.size
+    } catch(_) {
+      return null
+    }
+  }
+
+  /**
+   * Gets the last modification time of a file.
+   * Used by the caching system to determine if cached data is still valid.
+   *
+   * @returns {Promise<Date?>} The last modification time, or null if file doesn't exist
+   */
+  async modified() {
+    try {
+      const stat = await fs.stat(this.path)
+
+      return stat.mtime
+    } catch(_) {
+      return null
+    }
+  }
+
+  /**
+   * @typedef {object} FileParts
+   * @property {string} base - The file name with extension
+   * @property {string} dir - The directory path
+   * @property {string} ext - The file extension (including dot)
+   */
+
+  /**
+   * Deconstruct a filename into parts
+   *
+   * @param {string} fileName - The filename to deconstruct
+   * @returns {FileParts} The filename parts
+   */
+  #deconstructFilenameToParts(fileName) {
+    Valid.assert(typeof fileName === "string" && fileName.length > 0,
+      "file must be a non-zero length string", 1)
+
+    return path.parse(fileName)
+  }
+
+  /**
+   * Reads the content of a file asynchronously.
+   *
+   * @param {string} [encoding] - The encoding to read the file as.
+   * @returns {Promise<string>} The file contents
+   */
+  async read(encoding="utf8") {
+    const filePath = this.path
+
+    if(!(await this.exists))
+      throw Sass.new(`No such file '${filePath}'`)
+
+    if(!filePath)
+      throw Sass.new("No absolute path in file map")
+
+    return await fs.readFile(filePath, encoding)
+  }
+
+  /**
+   * Writes content to a file synchronously.
+   *
+   * @param {string} content - The content to write
+   * @param {string} encoding - The encoding in which to write.
+   * @returns {Promise<void>}
+   */
+  async write(content, encoding="utf8") {
+    if(!this.path)
+      throw Sass.new("No absolute path in file")
+
+    await fs.writeFile(this.path, content, encoding)
+  }
+
+  /**
+   * Loads an object from JSON or YAML provided a fileMap
+   *
+   * @param {string} [type] - The expected type of data to parse.
+   * @param {string} [encoding] - The encoding to read the file as.
+   * @returns {object} The parsed data object.
+   */
+  async loadData(type="any", encoding="utf8") {
+    const content = await this.read(encoding)
+    const toTry = {
+      json5: [JSON5],
+      json: [JSON5],
+      yaml: [YAML],
+      any: [JSON5,YAML]
+    }[type.toLowerCase()]
+
+    for(const [format] of toTry) {
+      try {
+        const result = format.parse(content)
+
+        return result
+      } catch {
+        // nothing to see here
+      }
+    }
+
+    throw Sass.new(`Content is neither valid JSON5 nor valid YAML:\n'${this.path}'`)
   }
 }
